feat: support activitystream 2 simultaneously
We've (Data Infrastructure) deployed a new version of Activity Stream
running Elasticsearch 7, which has a slightly updated API. The 'total'
key is an object in ES7 and a plain int in ES6. We want to support both
activity streams for the transition so that we don't need to sync the
config change exactly.