feature: remove Aventri events from the search ElasticSearch query
The ElasticSearch/Activity Stream feed for Aventri is currently disabled but will be re-enabled soon. However once it is re-enabled it will return all events from the Aventri API, in contrast to the original feed which did some filtering. This will cause undesired search results to appear on great.gov.uk if the ElasticSearch query remains as it is. Therefore this commit removes Aventri events from the ElasticSearch query and a separate commit will be pushed at a later stage to update the query with the relevant filters.